De-duplicate provisioning code (#81)
