+++ libsignal-service/src/provisioning/mod.rs
@@ -1,0 +1,32 @@
+mod cipher;
+mod manager;
+mod pipe;
+
+pub use cipher::ProvisioningCipher;
+pub use manager::{
+    ConfirmCodeMessage, ConfirmDeviceMessage, LinkingManager,
+    ProvisioningManager, SecondaryDeviceProvisioning,
+};
+
+use crate::prelude::ServiceError;
+pub use crate::proto::{
+    ProvisionEnvelope, ProvisionMessage, ProvisioningVersion,
+};
+
+#[derive(thiserror::Error, Debug)]
+pub enum ProvisioningError {
+    #[error("Invalid provisioning data: {reason}")]
+    InvalidData { reason: String },
+    #[error("Protobuf decoding error: {0}")]
+    DecodeError(#[from] prost::DecodeError),
+    #[error("Websocket error: {reason}")]
+    WsError { reason: String },
+    #[error("Websocket closing: {reason}")]
+    WsClosing { reason: String },
+    #[error("Service error: {0}")]
+    ServiceError(#[from] ServiceError),
+    #[error("libsignal-protocol error: {0}")]
+    ProtocolError(#[from] libsignal_protocol::Error),
+    #[error("ProvisioningCipher in encrypt-only mode")]
+    EncryptOnlyProvisioningCipher,
+}
